Refactor extensions to TS
